// autogpt_platform/frontend/src/app/api/__generated__/models/searchEntryFilterAnyOfItem.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */

export type SearchEntryFilterAnyOfItem =
  (typeof SearchEntryFilterAnyOfItem)[keyof typeof SearchEntryFilterAnyOfItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SearchEntryFilterAnyOfItem = {
  blocks: "blocks",
  integrations: "integrations",
  marketplace_agents: "marketplace_agents",
  my_agents: "my_agents",
} as const;